Move Interactor to correct package; move interactor down into ViewModel instead of being in Fragment
